+++ examples/advanced/gnn-pt/jobs/gnn-pt/app/config/config_fed_server.json
