+++ examples/advanced/gnn-pt/jobs/gnn-pt/app/custom/graphsage_trainer.py
@@ -36,9 +36,6 @@ from torch_geometric.nn import GraphSAGE
 import tqdm
 
 
-
-
-
 class GraphSageTrainer(Executor):
     def __init__(
         self,
