+++ examples/advanced/gnn-pt/jobs/gnn-pt/app/custom/graphsage_validator.py
