+++ examples/advanced/gnn-pt/jobs/gnn-pt/app/custom/simple_network.py
